backend/app/services/synthesis.py
import ffmpeg
import os
from app.core.config import settings

def render_karaoke_video(assets: dict) -> str:
    """
    Combines background video, instrumental audio, and subtitles into a final video.
    """
    instrumental_path = assets.get("instrumental")
    background_path = assets.get("background")
    subtitles_path = assets.get("subtitles") # ASS file path
    output_path = assets.get("output_path", os.path.join(settings.TEMP_DIR, "output.mp4"))

    print(f"Mock: Rendering video to {output_path}")

    # Real implementation using ffmpeg-python
    # input_video = ffmpeg.input(background_path, stream_loop=-1)
    # input_audio = ffmpeg.input(instrumental_path)

    # (Optional: Filter complex for resizing background)

    # stream = ffmpeg.output(
    #     input_video,
    #     input_audio,
    #     output_path,
    #     vf=f"ass={subtitles_path}",
    #     shortest=None,
    #     vcodec='libx264',
    #     acodec='aac'
    # )
    # ffmpeg.run(stream, overwrite_output=True)

    return output_path
