backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.1
requests==2.31.0
yt-dlp==2023.11.16
demucs==4.0.1
torch==2.1.2
torchaudio==2.1.2
ffmpeg-python==0.2.0
# WhisperX needs to be installed specifically, often via git, but for requirements file we can try this or rely on Docker/pip install
git+https://github.com/m-bain/whisperX.git
google-generativeai==0.3.2
pydantic==2.6.1
pydantic-settings==2.1.0
